Refactor: CommentBottomSheet 리팩토링
