+++ src/components/CommentBottomSheet/Comment/styles.tsx
