+++ src/components/CommentBottomSheet/styles.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import CloseIcon from '../../assets/default/modal-close-white.svg';
 
 export const CommentBottomSheetLayout = styled.div`
 	${({ theme }) => theme.visibleOnMobileTablet};
@@ -54,13 +53,11 @@ export const CommentModalBox = styled.section`
 	width: 100%;
 `;
 
-export const XButton = styled.button`
+export const CloseButton = styled.button`
 	width: 1.875rem;
 	height: 1.875rem;
-	margin: auto 0 auto auto;
-	background-image: url(${CloseIcon});
-	background-repeat: no-repeat;
-	background-size: 1.875rem;
-	background-position: center;
+	display: flex;
+	justify-content: center;
+	align-items: center;
 	opacity: 0.5;
 `;
